refactor: move errors which are unused in libmrot to mrot app
